process-image/requirements.txt
Flask
rembg
Pillow
python-dotenv
boto3